// src/generated/types.ts
export interface AmountChangeDto {
  comments?: string | null
  amount: number
  cancelDate?: string | null
  creationDate?: string | null
  cancelComments?: string | null
  changedId: number
  id: number
}

export interface AssignmentDto {
  assignmentContractNr?: string | null
  assignmentDebtorName?: string | null
  assignmentDebtorCompanyNr?: string | null
  assignmentIdfNr?: string | null
  originalIdfNr?: string | null
}

export interface CancelInvoiceMessageDto {
  invoiceId: number
  cancelComments?: string | null
}

export interface CancellationDto {
  cancelDate?: string | null
  cancelledBy?: number | null
  reason?: string | null
}

export interface ChangeStatusInvoiceResponse {
  invoiceNumber?: string | null
  status: PurchaseInvoiceStatus
}

export interface ChangeUser {
  firstName?: string | null
  lastName?: string | null
}

export interface CirHistoryDto {
  comment?: string | null
  assignment: AssignmentDto
  amountChanges?: AmountChangeDto[] | null
  cancellation: CancellationDto
  settlements?: SettlementDto[] | null
}
// ...
export interface CompanyDto {
  companyId: number
  countryId: number
  name: string
  webAddress?: string | null
  addresses?: SspAddressDto[] | null
  bankAccounts?: SspBankAccountDto[] | null
  emails?: SspEmailDto[] | null
  registrationCode: string
  vatRegistrationCode?: string | null
  phoneNumber?: string | null
  contactPerson?: string | null
  contactEmail?: string | null
  logo?: string | null
  settings: CompanySettingsDto
  isMainCompany: boolean
  groupId?: number[] | null
  isPrivateCompany: boolean
  packageId: number
  additionalCode?: string | null
  plusChannelsActive: boolean
  status: CompanyStatus
  companyWillBeDeletedAt?: string | null
  serbiaCompanyType: SerbiaCompanyType
  nonSebIbanWarning: boolean
  storeInvoiceDetails?: boolean | null
  hasISP?: boolean | null
  informationServiceProviderId?: number | null
  informationServiceProviderName?: string | null
  ispAcceptedToRepresentCompany?: boolean | null
  specialTreatment: boolean
  budgetCompanyType?: string | null
}

export interface CompanyFullDetailsDto {
  id: number
  vatRegistrationCode?: string | null
  name?: string | null
  address?: string | null
  registrationCode?: string | null
  isBudget: boolean
}

export interface CompanySettingsDto {
  homeRoute?: string | null
  modules?: SspModules[] | null
}

export interface CompanyUpdateNotificationEndpointDto {
  outgoingInvoicesEndpoint?: string | null
  incomingInvoicesEndpoint?: string | null
  incomingInvoicesPPCEndpoint?: string | null
}

export interface ContractApplicationDto {
  contractApplicationId: number
  companyName?: string | null
  registrationCode?: string | null
  vatRegistrationCode?: string | null
  serviceId?: string | null
  addresses?: SspAddressDto[] | null
  channel?: number | null
  channelAddress?: string | null
  invoiceNotificationEmail?: string | null
  serviceProvider?: string | null
  paymentTerms?: number | null
  fineRatePerDay?: number | null
  referenceNr?: string | null
  contractDate: string
  companyLabel?: string | null
  companyEmail?: string | null
  companyPhone?: string | null
  contactPersonName?: string | null
  contactPersonEmail?: string | null
  receiverCompanyId: number
}

export interface ContractCustomFieldDto {
  id: number
  name?: string | null
  value?: string | null
}
// ...
export interface DiscountDto {
  amount: number
  vatRate: number
  vatCategoryCode?: string | null
}

export interface EfakturaVersionDto {
  version?: string | null
}

export interface EFiscalizationFiscalBillNumbersForInvoiceRepresentation {
  fiscalBillNumber?: string | null
  transactionType: TransactionType
  sufTaxCode?: string | null
}

export interface ExemptionReasonByVatCategoryDto {
  vatCategoryCode?: string | null
  reasonCode?: string | null
  exemptionDecisionNumber?: string | null
  feeItemsValueWithoutVat: number
  discountFeeWithoutVat: number
  feeWithoutVat: number
}
// ...
export interface InvoiceChangeDto {
  id: number
  propertyName?: string | null
  oldValue?: string | null
  newValue?: string | null
  dateChanged: string
  user: ChangeUser
  version: number
  serviceDesk: boolean
  ispName?: string | null
}

export interface InvoiceDto {
  serviceId?: string | null
  invoiceId: number
  senderId: number
  sender?: string | null
  receiver?: string | null
  invoiceNumber?: string | null
  accountingDateUtc?: string | null
  paymentDateUtc?: string | null
  invoiceDateUtc?: string | null
  invoiceSentDateUtc?: string | null
  referenceNumber?: string | null
  fineRatePerDay: number
  description?: string | null
  note?: string | null
  orderNumber?: string | null
  currency?: string | null
  discountPercentage: number
  discountAmount: number
  sumWithoutVat: number
  vatRate: number
  vatSum: number
  sumWithVat: number
  createdUtc: string
  lastModifiedUtc: string
  version: number
  modelNumber?: string | null
  receiverId?: number | null
  publicPurchaseContractSigner: SspCustomerSupplierDto
  publicPurchaseContractSignerContractId?: number | null
  contract: SspContractDto
  status: SalesInvoiceStatus
  uasSender: CompanyDto
  totalRowsCount: number
  channel?: number | null
  channelAddress?: string | null
  serviceProvider?: string | null
  rows?: LineItemDto[] | null
  invoiceMessage?: string | null
  acceptRejectMessage?: string | null
  attachments?: SalesInvoiceAttachmentDto[] | null
  bankAccounts?: SalesInvoiceBankAccountDto[] | null
  selectedPrepaymentInvoices?: SelectedIndividualPrepaymentInvoiceDto[] | null
  isCreditInvoice: boolean
  senderReceiverContractNumber?: string | null
  errorCode: ErrorCodes
  totalToPay?: number | null
  sendInvoiceToCir?: boolean | null
  cirInvoiceId?: string | null
  cirAmountChangeId?: string | null
  cirSettledAmount: number
  isProFormaInvoice?: boolean | null
  cirHistory: CirHistoryDto
  cirAssignationHistory: InvoiceHistoryDto
  cirStatus: CirInvoiceStatus
  isDebitNote: boolean
  stornoNumber?: string | null
  cancelInvoiceMessage?: string | null
  prepaymentInvoiceNumber?: string | null
  isPrepaymentInvoice: boolean
  vatPointDate: number
  globUniqId?: string | null
  vatNumberFactoringCompany?: string | null
  factoringContractNumber?: string | null
  sourceInvoiceSelectionMode: SourceInvoiceSelectionMode
  indebtednessPeriodFromDate?: string | null
  indebtednessPeriodToDate?: string | null
  sourceInvoices?: InvoiceLinkDto[] | null
  creditInvoices?: InvoiceLinkDto[] | null
  debitNotes?: InvoiceLinkDto[] | null
  prepaymentCalculation: SalesPrepaymentCalculationDto
  prepaymentCalculationDomestic: SalesPrepaymentCalculationDto
  invoiceTotalPaymentsCalculation: SalesInvoiceTotalPaymentsCalculationDto
  invoiceTotalPaymentsCalculationDomestic: SalesInvoiceTotalPaymentsCalculationDto
  lotNumber?: string | null
  exemptionReasonsByVatCategory?: ExemptionReasonByVatCategoryDto[] | null
  discounts?: DiscountDto[] | null
  createdFromXml?: boolean | null
  roundingAmount: number
  fiscalBillNumbers?: string[] | null
  fiscalBillNumbersRepresentation?: EFiscalizationFiscalBillNumbersForInvoiceRepresentation[] | null
}

export interface InvoiceHistoryDto {
  invoiceId: number
  invoiceChanges?: InvoiceChangeDto[] | null
}

export interface InvoiceLinkDto {
  invoiceId?: string | null
  cirInvoiceId?: string | null
  invoiceNumber?: string | null
  sentToCir: boolean
  status: SalesInvoiceStatus
}

export interface InvoiceTotalVatPaymentsCalculationDomesticDto {
  vatPerRateCalculations?: InvoiceTotalVatPerRateCalculationDomesticDto[] | null
  totalVatPaymentsDomestic: number
}

export interface InvoiceTotalVatPerRateCalculationDomesticDto {
  vatRate: number
  baseSumForPaymentVatRate: number
  vatCategory?: string | null
  vatPaymentPerRate: number
}

export interface LineItemDto {
  rowId: number
  invoiceId: number
  orderNo: number
  code?: string | null
  description?: string | null
  unit?: string | null
  unitPrice: number
  quantity: number
  discountPercentage: number
  discountAmount: number
  sumWithoutVat: number
  vatRate: number
  vatSum: number
  sumWithVat: number
  vatCategoryCode?: string | null
  classificationCode?: string | null
  allowances?: SalesLineItemAllowanceDto[] | null
  despatchAdviceId?: string | null
  receiptAdviceId?: string | null
}

export interface MiniCompanyDto {
  bugetCompanyNumber?: string | null
  registrationCode?: string | null
  vatRegistrationCode?: string | null
  name?: string | null
  registrationDate?: string | null
  deletionDate?: string | null
}

export interface MiniInvoiceDto {
  invoiceId: number
  purchaseInvoiceId: number
  salesInvoiceId: number
}
// ...
export interface SalesInvoiceAttachmentDto {
  invoiceId?: number | null
  fileName?: string | null
  fileData?: string | null
  id: number
  fileSize: number
  createdUtc: string
  isUbl: boolean
  isLink: boolean
  link?: string | null
}

export interface SalesInvoiceBankAccountDto {
  bankAccountId: number
  salesInvoiceId: number
  unifiedBankAccount?: string | null
}

export interface SalesInvoicesDto {
  salesInvoiceIds?: number[] | null
}

export interface SalesInvoiceStatusChangeDto {
  eventId: number
  date?: string | null
  newInvoiceStatus: SalesInvoiceStatus
  salesInvoiceId: number
  comment?: string | null
  cirInvoiceId?: string | null
  subscriptionKey?: string | null
  stornoNumber?: string | null
  cirAssignmentChange: CirAssignmentChange
  isSigned?: boolean | null
  isAutoAssigned: boolean
}

export interface SalesInvoiceTotalPaymentsCalculationDto {
  totalPayments: number
  paymentFeeForVatCategories?: VatCategoryDetailsDto[] | null
}

export interface SalesLineItemAllowanceDto {
  salesLineItemAllowanceId: number
  salesLineItemId: number
  baseAmount?: number | null
  percentage?: number | null
  amount: number
  reasonCode?: string | null
  reason?: string | null
}

export interface SalesPrepaymentCalculationDto {
  totalPrepaymentWithoutVat: number
  prepaymentPaidVat: number
  totalPrepaymentInvoice: number
  prepaymentFeeReductionForVatCategories?: VatCategoryDetailsDto[] | null
}
// ...
export interface SelectedIndividualPrepaymentInvoiceDto {
  selectedPrepaymentInvoiceId: number
  prepaymentInvoiceId: number
  prepaymentInvoiceNumber?: string | null
  prepayedAmount: number
  prepayedVAT: number
  currency?: string | null
  paymentDate?: string | null
  contractId: number
  invoiceSentDateUtc?: string | null
  individualPrepaymentCalculations?: SalesIndividualPrepaymentCalculationDto[] | null
}
// ...
export interface SettlementDto {
  settlementDate?: string | null
  amount: number
  comment?: string | null
}
// ...
export interface SimpleSalesInvoiceDto {
  invoiceId: number
  globUniqId?: string | null
  comment?: string | null
  cirStatus: CirInvoiceStatus
  cirInvoiceId?: string | null
  version: number
  lastModifiedUtc: string
  cirSettledAmount: number
  vatNumberFactoringCompany?: string | null
  factoringContractNumber?: string | null
  cancelComment?: string | null
  stornoComment?: string | null
  status: SalesInvoiceStatus
}
// ...
export interface SspAddressDto {
  addressId: number
  streetAndHouse?: string | null
  postalIndex?: string | null
  city?: string | null
  countryId?: number | null
}

export interface SspBankAccountDto {
  bankAccountId: number
  bankId?: number | null
  currency?: string | null
  iban?: string | null
  swift?: string | null
  isPrimary: boolean
  unifiedBankAccount?: string | null
}

export interface SspContractDto {
  contractId: number
  customerSupplier: SspCustomerSupplierDto
  customerSupplierId?: number | null
  paymentTerms?: number | null
  fineRatePerDay?: number | null
  serviceId?: string | null
  referenceNr?: string | null
  contactPersonName?: string | null
  contactPersonMail?: string | null
  contractDesc?: string | null
  addresses?: SspAddressDto[] | null
  companyAddress: SspAddressDto
  customFields?: ContractCustomFieldDto[] | null
  channels?: number[] | null
  isDeleted: boolean
  ownerCompanyId?: number | null
  channel?: number | null
  channelAddress?: string | null
  serviceProvider?: string | null
  contractNumber?: number | null
  invoiceNotificationEmail?: string | null
  invoiceCurrency?: string | null
  isBudgetClient: boolean
  sendInvoiceToCir: boolean
  availableContracts?: string[] | null
  contractApplication: ContractApplicationDto
  isReceiverCompanyDeleted: boolean
}

export interface SspCustomerSupplierDto {
  customerSupplierId: number
  ownerCompanyId?: number | null
  countryId?: number | null
  companyName?: string | null
  registrationCode?: string | null
  vatRegistrationCode?: string | null
  firstName?: string | null
  lastName?: string | null
  personalId?: string | null
  phoneNumber?: string | null
  isCompany?: boolean | null
  companyMail?: string | null
  email?: string | null
  language: Languages
  additionalCode?: string | null
  serbiaCompanyType: SerbiaCompanyType
  specialTreatment: boolean
}

export interface SspEmailDto {
  email: string
  isActivated: boolean
  activationToken?: string | null
}
// ...
export interface ValueAddedTaxExemptionReasonDto {
  reasonId: number
  key?: string | null
  law?: string | null
  article?: string | null
  paragraph?: string | null
  point?: string | null
  subpoint?: string | null
  text?: string | null
  freeFormNote?: string | null
  activeFrom: string
  activeTo?: string | null
  category?: string | null
}